==> 7_part5_final.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("TraditionMacroFactors.csv")

ret = pd.read_csv("portfolio_return.csv").set_index("Unnamed: 0")
port_ret = ret.sum(axis = 1)

port_ret = port_ret.to_frame()

port_ret.columns = ['Return']

df_merge = pd.merge(df, port_ret, on='Unnamed: 0')
correlation_sherrill = df_merge.corr()

#df_merge.rename(columns = {'0':'Return'}, inplace = True)

correlation_sherrill[6:7]

ret1 = pd.read_csv("portfolio1.csv").set_index("date")
port_ret1 = ret1.sum(axis = 1)

ret1.iloc[1:51].sum(axis=1)

.set_index("Unnamed: 0")
